app/requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==23.2.0
autoflake==2.2.1
autopep8==2.0.4
beautifulsoup4==4.12.3
bleach==6.1.0
soupsieve==2.5
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
django-parler==2.3
django-parler-rest==2.2
django-nested-admin==4.0.2
drf-jwt==1.19.2
drf-nested-routers==0.93.5
drf-yasg==1.21.7
drf-writable-nested==0.7.0
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.13.2
lazy-object-proxy==1.10.0
Markdown==3.5.2
mccabe==0.7.0
packaging==23.2
pluggy==1.4.0
psycopg2==2.9.9
py==1.11.0
pycodestyle==2.11.1
pylint==2.17.7
pyparsing==3.1.1
pytest==7.4.1
pytest-drf==1.1.3
pytest-django==4.8.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-randomly==3.15
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
wrapt==1.16.0
whitenoise==6.6.0
dadata==21.10.1
pexpect==4.9.0
pur==7.3.1
pandas==2.2.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pygsheets==2.0.6
django-spaghetti-and-meatballs==0.4.2
Unidecode==1.3.8
sentry-sdk==1.40.4
django-simple-history==3.4.0
openpyxl==3.1.2
xlrd==2.0.1
html2text==2020.1.16
six==1.16.0
webencodings==0.5.1
redbox==0.2.1
coreapi==2.3.3
flake8==6.1.0
flake8-pytest==1.4
flake8-pytest-style==1.7.2
flake8_simplify==0.21.0
flake8-variables-names==0.0.6
flake8-mutable==1.2.0
flake8-django==1.4
flake8-cognitive-complexity==0.1.0
flake8-bugbear==24.2.6
flake8-print==5.0.0
flake8-multiline-containers==0.0.19
flake8-pie==0.16.0
pep8-naming==0.13.3
mixer==7.2.2
mypy==1.8.0
djangorestframework-stubs==3.14.4
django-stubs==4.2.6
django-health-check==3.17
black==24.2.0
mimesis==11.1.0
django-phonenumber-field[phonenumberslite]==7.3.0
